+++ app/User.php
@@ -4,28 +4,23 @@ namespace App;
 
 use Illuminate\Foundation\Auth\User as Authenticatable;
 
-
 class User extends Authenticatable
 {
-    public static $roles = ['admin', 'editor'];
+    public static $roles = ['admin', 'editor', 'importer'];
 
-    protected $fillable = [
-        'name', 'email', 'role', 'username'
-    ];
+    protected $fillable = ['name', 'email', 'role', 'username'];
 
     /**
      * The attributes that should be hidden for arrays.
      *
      * @var array
      */
-    protected $hidden = [
-        'password', 'remember_token',
-    ];
+    protected $hidden = ['password', 'remember_token'];
 
-    public static $rules = array(
+    public static $rules = [
         'name' => 'required',
-        'role' => 'in:admin,editor|required',
+        'role' => 'in:admin,editor,importer|required',
         'email' => 'email|required',
         'username' => 'required',
-    );
+    ];
 }