+++ database/migrations/2022_08_19_124213_add_indexes_to_item_images.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration {
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('item_images', function (Blueprint $table) {
+            $table->index(['item_id']);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('item_images', function (Blueprint $table) {
+            $table->dropIndex(['item_id']);
+        });
+    }
+};