+++ resources/views/authorities/form.blade.php
@@ -30,7 +30,7 @@
     <div class="panel-body">
       <div class="row">
 
-        @if(isset($new_id))     
+        @if(isset($new_id))
           <div class="col-md-12">
             <div class="form-group">
               {!! Form::label('id', 'Id') !!}
@@ -62,7 +62,7 @@
 
               <div class="form-group">
                 {{ Form::label($locale . "[biography]", 'Biografia '.strtoupper($locale)) }}
-                {{ Form::textarea($locale . "[biography]", isset($authority) ? @$authority->translate($locale)->biography : '', array('class' => 'form-control wysiwyg', 'rows'=>'12')) }}  
+                {{ Form::textarea($locale . "[biography]", isset($authority) ? @$authority->translate($locale)->biography : '', array('class' => 'form-control wysiwyg', 'rows'=>'12')) }}
               </div>
 
             </div>
@@ -86,7 +86,7 @@
       Použité zdroje
     </div>
     <div class="panel-body">
-      <admin-links-input 
+      <admin-links-input
         collection-name="sourceLinks"
         :value="{{ json_encode(Request::old('sourceLinks', $authority->sourceLinks)) }}"
       ></admin-links-input>
@@ -100,7 +100,7 @@
       Externé odkazy
     </div>
     <div class="panel-body">
-      <admin-links-input 
+      <admin-links-input
         collection-name="externalLinks"
         :value="{{ json_encode(Request::old('externalLinks', $authority->externalLinks)) }}"
       ></admin-links-input>
@@ -136,20 +136,26 @@
             <div class="cropit-image-preview-container">
               <div class="cropit-image-preview"></div>
             </div>
-            
+
             <div class="image-size-label">&nbsp;</div>
             <div class="form-group" style="padding: 0 15px">
               <input type="text" class="cropit-image-zoom-input" min="0" max="1" step="0.01" data-slider-min="0" data-slider-max="1" data-slider-step="0.01" data-slider-value="0">
             </div>
 
             <input type="file" class="cropit-image-input" />
-            <a class="btn btn-success btn-outline select-image-btn"><i class="fa fa-picture-o"></i> nahrať obrázok</a>
-            {!! Form::hidden('primary_image', null, ['id' => 'primary_image']) !!}
-            
           </div>
+
+          <a class="btn btn-success btn-outline select-image-btn"><i class="fa fa-picture-o"></i> nahrať obrázok</a>
+          <br />
+          <button type="button" id="clear-image" class="mt-3 btn btn-danger btn-xs btn-outline" style="display:none">
+            <i class="fa fa-times"></i> zmazať obrázok
+          </button>
+
+          {!! Form::hidden('primary_image', null, ['id' => 'primary_image']) !!}
+          {!! Form::hidden('clear_primary_image', null, ['id' => 'clear_primary_image']) !!}
         </div>
 
-      </div> 
+      </div>
       <!-- /.row (nested) -->
     </div>
     <!-- /.panel-body -->
@@ -158,7 +164,7 @@
 </div>
 
 <div class="col-md-12 text-center">
-  {!! Form::submit('Uložiť', array('class' => 'btn btn-default')) !!} &nbsp; 
+  {!! Form::submit('Uložiť', array('class' => 'btn btn-default')) !!} &nbsp;
   @if(isset($authority) && $authority->record)
     <a href="{!!URL::to('harvests/'.$authority->record->id.'/refreshRecord')!!}" class="btn btn-warning">Obnoviť z OAI</a>
   @endif
@@ -176,13 +182,21 @@
 
 <script>
 $(document).ready(function(){
+  const $clearImageButton = $('button#clear-image');
+  const $clearImageInput = $('input#clear_primary_image')
+
   $(".cropit-image-zoom-input").slider({
     tooltip: 'hide'
   });
 
   $('#image-editor').cropit({
     imageBackground: true,
-    imageBackgroundBorderWidth: 20
+    imageBackgroundBorderWidth: 20,
+    onImageLoaded: () => {
+      $('#image-editor').show();
+      $clearImageButton.show();
+      $clearImageInput.val('')
+    }
     @if (isset($authority) && $authority->has_image && file_exists($authority->getImagePath(true)) )
       ,imageState: {
         src: '{!! $authority->getImagePath() !!}'
@@ -194,6 +208,12 @@ $(document).ready(function(){
     $('.cropit-image-input').click();
   });
 
+  $clearImageButton.click(function() {
+    $('#image-editor').hide()
+    $clearImageButton.hide()
+    $clearImageInput.val(true)
+  });
+
   $('form').submit(function(e) {
     var imageData = $('#image-editor').cropit('export', {
       type: 'image/jpeg',
@@ -202,7 +222,7 @@ $(document).ready(function(){
     $('#primary_image').val(imageData);
       return true;
   });
-  
+
 });
 
 </script>