+++ tests/Importers/WebumeniaMgImporterTest.php
@@ -8,6 +8,7 @@ use App\Importers\WebumeniaMgImporter;
 use App\Matchers\AuthorityMatcher;
 use Illuminate\Contracts\Translation\Translator;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use SplFileInfo;
 use Tests\TestCase;
 
 class WebumeniaMgImporterTest extends TestCase
@@ -169,8 +170,7 @@ class WebumeniaMgImporterTest extends TestCase
             $this->app->get(Translator::class)
         );
         $import = Import::create();
-        $file = ['basename' => '', 'path' => ''];
-        $items = $importer->import($import, $file);
+        $items = $importer->import($import, new SplFileInfo(''));
 
         $this->assertEquals('', $import->records()->first()->error_message);
         $this->assertCount(1, $items);