+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.64.0] - 2022-08-19
+### Fixed
+- item_images indexes
+
+### Added
+- importer role
+- import permissions
+- clear image functionality for authorities in admin
+- filter items by collection in admin
+- confirm collection delete
+
 ## [2.63.0] - 2022-08-19
 ### Added
 - item sortables